+++ packages/design-system-shared/src/index.ts
@@ -34,11 +34,28 @@ export { type TextOrChildrenPropsShared } from './types/TextOrChildren';
 // TitleHub types (ADR-0004)
 export { type TitleHubPropsShared } from './types/TitleHub';
 
-// BoxHorizontal types (ADR-0004)
-export { type BoxHorizontalPropsShared } from './types/BoxHorizontal';
+// BoxColumn types (ADR-0004)
+export { type BoxColumnPropsShared } from './types/BoxColumn';
 
-// BoxVertical types (ADR-0004)
-export { type BoxVerticalPropsShared } from './types/BoxVertical';
+// BoxRow types (ADR-0004)
+export { type BoxRowPropsShared } from './types/BoxRow';
+
+// HeaderSearch types (ADR-0003 + ADR-0004)
+export {
+  HeaderSearchVariant,
+  type HeaderSearchInlinePropsShared,
+  type HeaderSearchPropsShared,
+  type HeaderSearchScreenPropsShared,
+} from './types/HeaderSearch';
+
+// KeyValueColumn types (ADR-0004)
+export { type KeyValueColumnPropsShared } from './types/KeyValueColumn';
+
+// KeyValueRow types (ADR-0003 + ADR-0004)
+export {
+  KeyValueRowVariant,
+  type KeyValueRowPropsShared,
+} from './types/KeyValueRow';
 
 // ButtonFilter types (ADR-0004)
 export { type ButtonFilterPropsShared } from './types/ButtonFilter';
@@ -57,6 +74,9 @@ export {
   BannerAlertSeverity,
   type BannerAlertPropsShared,
 } from './types/BannerAlert';
+
+// BadgeNetwork types (ADR-0004)
+export { type BadgeNetworkPropsShared } from './types/BadgeNetwork';
 
 // BadgeWrapper types (ADR-0003 + ADR-0004)
 export {